#595972 : completed support for builtin type extensions

+++ querydsl-apt/src/test/java/com/mysema/query/domain/Inheritance4Test.java
@@ -26,9 +26,9 @@ public class Inheritance4Test extends AbstractTest{
 
     @QueryEntity
     public class EntityWithNumber extends EntityWithComparable{
-        private Integer field;
+        private Long field;
 
-        public Integer getField() {
+        public Long getField() {
             return field;
         }
